Fix failing xgboost.dask tests with dask=2024.12.1
This adds a patch that fixes a failing test with older versions of Dask.
The test was failing with a cancelled Future error, because some part of
the task graph optimization step was incorrectly releasing some piece of
data that was needed to compute the result.

By persisting the result, we force it to stay in cluster memory.

This does have user-facing behavior, beyond the bugfix. The results are
persisted in cluster memory. If users were immediately reducing the
result or storing it to disk, they'll notice higher memory usage.

This will not leak memory indefinitely. Once the user releases all
their references to the result, the Futures on the cluster will be
garbage collected.